fix: several bugs
1. disable got with -fno-pic, hence our linker handles relocations
2. add parallel build support
3. update container image
4. generic makefile for utils both osx and linux support
5. add printf attribute and fix logging
6. remove assembly files, only use inline assembly
7. fix struct alignment between efi and kernel
8. fix struct definition at virtio network
9. prefer virtio kbd over ps/2
10. generic script for starting qemu efi hda
11. change into -O2 optimization
12. fix mov instruction at task switching

also we have bugs:
1. printf strange behavior, sometimes prints incorrectly
2. linker has page fault when logging is not at debug
3. network cannot send reply packets like arp, ping etc

+++ efi/console.c
@@ -94,7 +94,7 @@ size_t video_printf(char_t* fmt, ...) {
 				case 'd':
 #if ___BITS == 64
 					if(l_flag == 2) {
-						ival = va_arg(args, int128_t);
+						ival = va_arg(args, uint64_t);
 					} else if(l_flag == 1) {
 						ival = va_arg(args, int64_t);
 					}
@@ -132,7 +132,7 @@ size_t video_printf(char_t* fmt, ...) {
 				case 'u':
 #if ___BITS == 64
 					if(l_flag == 2) {
-						uval = va_arg(args, uint128_t);
+						uval = va_arg(args, uint64_t);
 					} else if(l_flag == 1) {
 						uval = va_arg(args, uint64_t);
 					}
@@ -167,7 +167,7 @@ size_t video_printf(char_t* fmt, ...) {
 				case 'h':
 #if ___BITS == 64
 					if(l_flag == 2) {
-						uval = va_arg(args, uint128_t);
+						uval = va_arg(args, uint64_t);
 					} else if(l_flag == 1) {
 						uval = va_arg(args, uint64_t);
 					}
